feat:stomp 메세지 일부 추가, adjustmentgame 기능 일부 추가

+++ Backend/tranvel/src/main/java/com/ssafy/tranvel/entity/RandomGame.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class RandomGame {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int miniGameCode;
+    private Long miniGameCode;
 
     @Column(length = 16, name = "GameName")
     private String gameName;
